ref(replay): Extract JumpButtons to its own component and put it on all lists
